requirements.txt
# Minimum safe versions (use '>=' to avoid installing insecure older releases).
# Review and update regularly; run `pip-audit` in CI to catch vulnerable packages.
fastapi>=0.116.1
uvicorn[standard]>=0.20.0

# Django and essential dependencies for CI/CD
django>=5.0.14
django-environ>=0.12.0

# Runtime dependencies (copied from pyproject)
requests>=2.32.4
setuptools>=78.1.1
beautifulsoup4>=4.11.0
alpha_vantage>=3.0.0
dash>=2.9.0
fredapi>=0.5.0
httpx>=0.24.0

# Database
sqlalchemy>=2.0.0
pymongo>=4.0.0
alembic>=1.11.0

# GUI/Web server
gunicorn>=20.1.0

# Finance specific
yfinance>=0.2.65
yahoofinancials>=1.20
stockdex>=1.2.0
financepy>=1.0.1

# Data analysis
pandas>=2.0.0
matplotlib>=3.6.0
numpy>=1.24.0
seaborn>=0.13.0
pandas-datareader>=0.10.0
quantstats>=0.0.76
scikit-learn>=1.2.0
statsmodels>=0.14.0
plotly>=5.13.0

# Testing and development tools
pytest>=6.0
flake8>=4.0.0
factory-boy>=3.3.0  # Test factories for Django models
requests-mock>=1.11.0  # HTTP request mocking for tests
memory-profiler>=0.61.0  # Memory profiling for performance tests

# Django dependencies for CI/CD
django-crispy-forms>=2.4
crispy-bootstrap5>=2025.6
django-allauth>=65.0.0
django-celery-beat>=2.8.0
djangorestframework>=3.16.0
django-cors-headers>=4.8.0
drf-spectacular>=0.28.0
django-compressor>=4.5.0
django-model-utils>=5.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter for Django
celery>=5.5.0
redis>=6.0.0  # Redis for caching and celery broker
django-redis>=6.0.0  # Django Redis integration
channels>=4.0.0  # WebSocket support for Django
channels-redis>=4.0.0  # Redis backend for Django Channels

# Optional: security tooling
pip-audit>=2.0.0

# Indirect dependencies (pinned by Snyk to avoid vulnerabilities)
# If you use extra packages in pyproject, keep this file in sync.
aiohttp>=3.12.14
anyio>=4.4.0
dnspython>=2.6.1
fonttools>=4.43.0
h11>=0.16.0
pillow>=10.3.0
protobuf>=4.25.8
urllib3>=2.5.0
werkzeug>=3.0.6
zipp>=3.19.1
sqlparse>=0.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
